+++ app/(root)/layout.tsx
@@ -1,0 +1,29 @@
+import LandingFooter from "@/components/footer/LandingFooter";
+import LandingNavbar from "@/components/header/LandingNavbar";
+import { Geist, Geist_Mono } from "next/font/google";
+
+const geistSans = Geist({
+  variable: "--font-geist-sans",
+  subsets: ["latin"],
+});
+
+const geistMono = Geist_Mono({
+  variable: "--font-geist-mono",
+  subsets: ["latin"],
+});
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <main
+      className={`${geistSans.variable} ${geistMono.variable} antialiased min-h-screen flex flex-col justify-between`}
+    >
+      <LandingNavbar />
+      {children}
+      <LandingFooter />
+    </main>
+  );
+}
