+++ app/dashboard/layout.tsx
@@ -1,0 +1,12 @@
+import type React from "react";
+import { Inter } from "next/font/google";
+
+const inter = Inter({ subsets: ["latin"] });
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return <main className={inter.className}>{children}</main>;
+}
