--- app/privacy-policy/loading.tsx
@@ -1,5 +1,0 @@
-import PrivacyPolicySkeleton from '../components/PrivacyPolicySkeleton';
-
-export default function PrivacyPolicyLoading() {
-  return <PrivacyPolicySkeleton />;
-}
