+++ components/skeletons/PrivacyPolicySkeleton.tsx
