+++ components/skeletons/TermsOfServiceSkeleton.tsx
