+++ components/skeletons/loading.tsx
@@ -1,0 +1,5 @@
+import PrivacyPolicySkeleton from "@/components/skeletons/PrivacyPolicySkeleton";
+
+export default function PrivacyPolicyLoading() {
+  return <PrivacyPolicySkeleton />;
+}
