+++ types/privacy-policy.ts
